Add Supabase Auth (#47)
* Auth - in progress.

* Auth: add delete account route; create profle, login, and library page; general cleanup.
