+++ components/Base/BaseCarousel.vue
@@ -3,5 +3,3 @@
     <slot></slot>
   </div>
 </template>
-
-<script setup lang="ts"></script>
