+++ middleware/auth.ts
@@ -1,0 +1,7 @@
+export default defineNuxtRouteMiddleware((to, from) => {
+  const user = useSupabaseUser();
+
+  if (!user.value) {
+    return navigateTo('/login');
+  }
+});
