+++ nuxt.config.ts
@@ -21,7 +21,7 @@ export default defineNuxtConfig({
     ApiKeyListenNotes: process.env.API_KEY_LISTEN_NOTES,
   },
   devtools: { enabled: true },
-  modules: ['@nuxt/ui', '@vueuse/nuxt', '@pinia/nuxt', '@nuxt/image'],
+  modules: ['@nuxt/ui', '@vueuse/nuxt', '@pinia/nuxt', '@nuxtjs/supabase'],
   alias: {
     '~': `${__dirname}`,
     '@': `${__dirname}`,
@@ -34,6 +34,7 @@ export default defineNuxtConfig({
   ],
   ui: {
     global: true,
+    safelistColors: ['primary', 'success', 'error', 'warn'],
     icons: ['heroicons', 'mdi'],
   },
   imports: {
@@ -47,5 +48,15 @@ export default defineNuxtConfig({
   },
   tailwindcss: {
     cssPath: '@/assets/style/main.scss',
+  },
+  supabase: {
+    redirect: false,
+    redirectOptions: { login: '/login', callback: '/profile', exclude: [] },
+    cookieName: 'supabase',
+    cookieOptions: {
+      maxAge: 60 * 60 * 8,
+      sameSite: 'lax',
+      secure: true,
+    },
   },
 });
