+++ package.json
@@ -14,6 +14,7 @@
   },
   "devDependencies": {
     "@nuxt/devtools": "latest",
+    "@nuxtjs/supabase": "^1.1.3",
     "@vueuse/core": "^10.5.0",
     "@vueuse/nuxt": "^10.5.0",
     "nuxt": "^3.8.0",
@@ -23,7 +24,6 @@
   },
   "dependencies": {
     "@iconify-json/mdi": "^1.1.55",
-    "@nuxt/image": "^1.0.0",
     "@nuxt/ui": "^2.9.0",
     "@pinia/nuxt": "^0.5.0",
     "@types/howler": "^2.2.10",
