+++ pages/podcast/[id].vue
@@ -1,8 +1,8 @@
 <template>
   <section v-if="!isLoading" class="pb-20">
-    <!-- If: Details -->
+    <!-- if: details -->
     <template v-if="podcastDetails">
-      <!-- Intro -->
+      <!-- intro -->
       <div class="min-h-[325px] bg-gradient-to-b from-slate-900 to-slate-800 py-16 text-body-inv">
         <div class="container">
           <div
@@ -18,16 +18,19 @@
               <p v-if="podcastDetails.description" class="mb-4">
                 {{ Utils.stripHTML(podcastDetails.description) }}
               </p>
-              <a
-                v-if="podcastDetails.website"
-                :href="podcastDetails.website"
-                target="_blank"
-                rel="noopener noreferrer nofollow"
-                class="text-link-inv text-sm"
-              >
-                <UIcon name="i-heroicons-link" />
-                Website
-              </a>
+              <div>
+                <a
+                  v-if="podcastDetails.website"
+                  :href="podcastDetails.website"
+                  target="_blank"
+                  rel="noopener noreferrer nofollow"
+                  class="text-link-inv text-sm"
+                >
+                  <UIcon name="i-heroicons-link" />
+                  Website
+                </a>
+              </div>
+              <SubscribeButton :id="podcastDetails.id" class="my-4" />
             </div>
           </div>
         </div>
@@ -48,7 +51,7 @@
       </div>
     </template>
 
-    <!-- Else: No Details Found -->
+    <!-- else: no details found -->
     <template>
       <div class="min-h-[325px] bg-gradient-to-b from-slate-900 to-slate-800 py-16 text-body-inv">
         <div class="container">
@@ -62,7 +65,7 @@
 <script setup lang="ts">
 import { useRoute } from 'vue-router';
 import { PodClientService } from '~/shared/podcast/api/services';
-import * as PodcastTypes from '~/shared/podcast/api/types/podcast-get.d';
+import type * as PodcastTypes from '~/shared/podcast/api/types/podcast-get.d';
 import * as Utils from '~/shared/utils';
 
 const isLoading = computed(() => getPodcastData.status.value === 'pending');
