+++ shared/podcast/api/services/index.ts
@@ -5,13 +5,14 @@ import * as SearchTypes from '~/shared/podcast/api/types/search-get';
 import * as BestPodcastsTypes from '~/shared/podcast/api/types/best-podcasts-get';
 
 // type AppUseFetchOptions<T> = Parameters<typeof useFetch<T>>[1];
+type FetchOptions = Parameters<typeof $fetch>[1];
 
 export class PodClientService {
   /**
    * Best Podcasts: get best podcasts details
    */
   getBestPodcasts<ResponseData extends BestPodcastsTypes.ServerResponse>(
-    fetchOptions: BestPodcastsTypes.RouteParams & Parameters<typeof $fetch>[1]
+    fetchOptions: BestPodcastsTypes.RouteParams & FetchOptions
   ) {
     return $fetch<ResponseData>('/api/podcast/best-podcasts', fetchOptions);
   }
@@ -20,14 +21,14 @@ export class PodClientService {
    * Search: get full search details
    */
   getSearchPodcasts<ResponseData extends SearchTypes.ServerResponsePodcast>(
-    fetchOptions: SearchTypes.RouteParams & Parameters<typeof $fetch>[1]
+    fetchOptions: SearchTypes.RouteParams & FetchOptions
   ) {
     fetchOptions.query.type = 'podcast';
     return $fetch<ResponseData>('/api/podcast/search', fetchOptions);
   }
 
   getSearchEpisodes<ResponseData extends SearchTypes.ServerResponseEpisode>(
-    fetchOptions: SearchTypes.RouteParams & Parameters<typeof $fetch>[1]
+    fetchOptions: SearchTypes.RouteParams & FetchOptions
   ) {
     fetchOptions.query.type = 'episode';
     return $fetch<ResponseData>('/api/podcast/search', fetchOptions);
@@ -36,18 +37,14 @@ export class PodClientService {
   /**
    * Genres: get details for genres
    */
-  getGenres<ResponseData extends GenresTypes.ServerResponse>(
-    fetchOptions?: GenresTypes.RouteParams & Parameters<typeof $fetch>[1]
-  ) {
+  getGenres<ResponseData extends GenresTypes.ServerResponse>(fetchOptions?: GenresTypes.RouteParams & FetchOptions) {
     return $fetch<ResponseData>('/api/podcast/genres', fetchOptions);
   }
 
   /**
    * Podcast: get podcast details and episodes
    */
-  getPodcast<ResponseData extends PodcastTypes.ServerResponse>(
-    fetchOptions: PodcastTypes.RouteParams & Parameters<typeof $fetch>[1]
-  ) {
+  getPodcast<ResponseData extends PodcastTypes.ServerResponse>(fetchOptions: PodcastTypes.RouteParams & FetchOptions) {
     return $fetch<ResponseData>('/api/podcast/podcast-single', fetchOptions);
   }
 
@@ -55,7 +52,7 @@ export class PodClientService {
    * Typeahead: get results for typeahead query
    */
   getTypeahead<ResponseData extends TypeaheadTypes.ServerResponse>(
-    fetchOptions: TypeaheadTypes.RouteGetParams & Parameters<typeof $fetch>[1]
+    fetchOptions: TypeaheadTypes.RouteGetParams & FetchOptions
   ) {
     return $fetch<ResponseData>('/api/podcast/typeahead', fetchOptions);
   }
