+++ test/mock-data/server/api/podcast/best-podcasts.ts
@@ -1,0 +1,1036 @@
+import type * as BestPodcastTypes from '~/shared/podcast/api/types/best-podcasts-get';
+
+export const mockBestPodcasts: BestPodcastTypes.ServerResponseRaw = {
+  id: 93,
+  name: 'Business',
+  total: 746,
+  has_next: true,
+  podcasts: [
+    {
+      id: '34beae8ad8fd4b299196f413b8270a30',
+      rss: 'https://feeds.feedburner.com/WorklifeWithAdamGrant',
+      type: 'episodic',
+      email: 'podcasts@ted.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9mZWVkcy5mZWVkYnVybmVyLmNvbS9Xb3JrbGlmZVdpdGhBZGFtR3JhbnQ=',
+        spotify_url: 'https://open.spotify.com/show/4eylg9GZJOVvUhTynt4jjA',
+        youtube_url: '',
+        linkedin_url: 'https://www.linkedin.com/in/adammgrant/',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'AdamMGrant',
+        facebook_handle: 'AdamMGrant',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image: 'https://production.listennotes.com/podcasts/worklife-with-adam-grant-ted-KgaXjFPEoVc.1400x1400.jpg',
+      title: 'WorkLife with Adam Grant',
+      country: 'United States',
+      website:
+        'https://www.ted.com/podcasts/worklife?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [67, 97, 111, 93],
+      itunes_id: 1346314086,
+      publisher: 'TED',
+      thumbnail: 'https://production.listennotes.com/podcasts/worklife-with-adam-grant-ted-KgaXjFPEoVc.300x300.jpg',
+      is_claimed: false,
+      description:
+        '\n      <p>You spend a quarter of your life at work. You should enjoy it! Organizational psychologist Adam Grant takes you inside the minds of some of the world’s most unusual professionals to discover the keys to a better work life. From learning how to love your rivals to harnessing the power of frustration, one thing’s for sure: You’ll never see your job the same way again. Produced in partnership with Transmitter Media.</p>\n    ',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 74,
+      total_episodes: 131,
+      listennotes_url: 'https://www.listennotes.com/c/34beae8ad8fd4b299196f413b8270a30/',
+      audio_length_sec: 2322,
+      explicit_content: false,
+      latest_episode_id: '7bc6a4aad20d4219b7c574e7419af605',
+      latest_pub_date_ms: 1692676800000,
+      earliest_pub_date_ms: 1518044524125,
+      update_frequency_hours: 167,
+      listen_score_global_rank: '0.01%',
+    },
+    {
+      id: '28ba59be5b8346589e910e24d4b3eed7',
+      rss: 'https://pultepodcast.libsyn.com/rss',
+      type: 'episodic',
+      email: 'Bill@pulte.org',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9wdWx0ZXBvZGNhc3QubGlic3luLmNvbS9yc3M=',
+        spotify_url: 'https://open.spotify.com/show/31g21O5kSlCstxSswwtPzh',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: '',
+        facebook_handle: '',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image: 'https://production.listennotes.com/podcasts/the-pulte-podcast-8PvlfCgcR_X-xBWa8_-4MTR.1400x1400.jpg',
+      title: 'The Pulte Podcast',
+      country: 'United States',
+      website: 'http://www.pulte.org?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [171, 93, 94, 67],
+      itunes_id: 1525585134,
+      publisher: 'Bill Pulte | Giving Money and Knowledge',
+      thumbnail: 'https://production.listennotes.com/podcasts/the-pulte-podcast-gqCft_ZYOIz-xBWa8_-4MTR.300x300.jpg',
+      is_claimed: false,
+      description:
+        "I'm giving away money, rent, food, and knowledge to people in need. I've built and sold 8 figure companies and now I want to help people.",
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 73,
+      total_episodes: 12,
+      listennotes_url: 'https://www.listennotes.com/c/28ba59be5b8346589e910e24d4b3eed7/',
+      audio_length_sec: 446,
+      explicit_content: false,
+      latest_episode_id: 'e26262d976694428bc1cc8c7af791d1b',
+      latest_pub_date_ms: 1621426832000,
+      earliest_pub_date_ms: 1596040778010,
+      update_frequency_hours: 202,
+      listen_score_global_rank: '0.01%',
+    },
+    {
+      id: '0d362b13399240de97602ef614acdcbc',
+      rss: 'https://feeds.megaphone.fm/startup',
+      type: 'episodic',
+      email: 'admin@gimletmedia.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://play.google.com/music/m/Ihs2lujac7unjyp2u7hp6ale7hq?t=StartUp_Podcast',
+        spotify_url: 'https://open.spotify.com/show/5CnDmMUG0S5bSSw612fs8C',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'podcaststartup',
+        facebook_handle: 'hearstartup',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image: 'https://production.listennotes.com/podcasts/startup-podcast-gimlet-14zU0c_MOmv-n9PpCBTQvoJ.1400x1400.jpg',
+      title: 'StartUp Podcast',
+      country: 'United States',
+      website: 'https://www.gimletmedia.com/startup?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [97, 93, 157, 94, 68, 127, 67, 171],
+      itunes_id: 913805339,
+      publisher: 'Gimlet',
+      thumbnail: 'https://production.listennotes.com/podcasts/startup-podcast-gimlet-8If7QBKU5jb-n9PpCBTQvoJ.300x300.jpg',
+      is_claimed: false,
+      description: "A series about what it's really like to start a business.",
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 73,
+      total_episodes: 153,
+      listennotes_url: 'https://www.listennotes.com/c/0d362b13399240de97602ef614acdcbc/',
+      audio_length_sec: 2176,
+      explicit_content: false,
+      latest_episode_id: '3663e1ba8f944df7956378ab332bf12b',
+      latest_pub_date_ms: 1598004000000,
+      earliest_pub_date_ms: 1396742400151,
+      update_frequency_hours: 253,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: '5f237b79824e4dfb8355f6dff9b1c542',
+      rss: 'https://feeds.npr.org/510325/podcast.xml',
+      type: 'episodic',
+      email: 'podcasts@npr.org',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9mZWVkcy5ucHIub3JnLzUxMDMyNS9wb2RjYXN0LnhtbA==',
+        spotify_url: 'https://open.spotify.com/show/4X3yDKgVTWRjSd6r0vhgo4',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'planetmoney',
+        facebook_handle: 'planetmoney',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/the-indicator-from-planet-money-npr-zjaIT2Ovo-r-G2EDjFO-TLA.1400x1400.jpg',
+      title: 'The Indicator from Planet Money',
+      country: 'United States',
+      website:
+        'https://www.npr.org/podcasts/510325/the-indicator-from-planet-money?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [99, 98, 144, 171, 93, 67],
+      itunes_id: 1320118593,
+      publisher: 'NPR',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/the-indicator-from-planet-money-npr-pb1-q5f_msG-G2EDjFO-TLA.300x300.jpg',
+      is_claimed: false,
+      description:
+        "A little show about big ideas. From the people who make <em>Planet Money</em>, <em>The Indicator</em> helps you make sense of what's happening today. It's a quick hit of insight into work, business, the economy, and everything else. Listen weekday afternoons.<br><br><em>Try Planet Money+! a new way to support the show you love, get a sponsor-free feed of the podcast, *and* get access to bonus content. You'll also get access to The Indicator and Planet Money Summer School, both without interruptions. sign up at plus.npr.org/planetmoney</em>",
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 73,
+      total_episodes: 1334,
+      listennotes_url: 'https://www.listennotes.com/c/5f237b79824e4dfb8355f6dff9b1c542/',
+      audio_length_sec: 573,
+      explicit_content: false,
+      latest_episode_id: '776c14d8e75242169b90fd8268c10325',
+      latest_pub_date_ms: 1694738992000,
+      earliest_pub_date_ms: 1527108300299,
+      update_frequency_hours: 29,
+      listen_score_global_rank: '0.01%',
+    },
+    {
+      id: 'a10d3bea3ffc40329727291321721c11',
+      rss: 'https://feeds.megaphone.fm/TPG1669083856',
+      type: 'episodic',
+      email: 'listen@thepodglomerate.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9mZWVkcy5tZWdhcGhvbmUuZm0vVFBHMTY2OTA4Mzg1Ng==',
+        spotify_url: '',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: '',
+        facebook_handle: '',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/money-rehab-with-nicole-lapin-money-news-ixsLZayFUCK-AvJIC3_u-j8.1400x1400.jpg',
+      title: 'Money Rehab with Nicole Lapin',
+      country: 'United States',
+      website: 'https://moneynewsnetwork.com/?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [93, 111, 181, 67],
+      itunes_id: 1559564016,
+      publisher: 'Money News Network',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/money-rehab-with-nicole-lapin-money-news-v24RloNcmpf-AvJIC3_u-j8.300x300.jpg',
+      is_claimed: false,
+      description:
+        'Ever notice that we will talk about everything before we talk about money? Sex? No problem. Politics? Bring it on. Money? Totally taboo. But not for long! Nicole Lapin— the only financial expert you don’t need a dictionary to understand, New York Times best selling author of Rich Bitch, and host of Money Rehab— is here to rehab your wallet, so you can get your financial life together once and for all. Episodes are just ten minutes-ish... no frills, just bite-sized tips and tricks so you don’t waste any time. And Nicole wants to hear from YOU! Email the money questions you want answered to MoneyRehab@NicoleLapin.com and Nicole will help–and you could even join Nicole on the show for a one-on-one intervention.',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 72,
+      total_episodes: 614,
+      listennotes_url: 'https://www.listennotes.com/c/a10d3bea3ffc40329727291321721c11/',
+      audio_length_sec: 969,
+      explicit_content: true,
+      latest_episode_id: 'c3c0786ac6ca46c89dc54053824c3de7',
+      latest_pub_date_ms: 1694761200000,
+      earliest_pub_date_ms: 1616593020570,
+      update_frequency_hours: 29,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: 'fc6d33e22b7f4db38df3bb64a9a8c227',
+      rss: 'https://tonyrobbins.libsyn.com/rss',
+      type: 'episodic',
+      email: 'tonyrobbins.social@gmail.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly90b255cm9iYmlucy5saWJzeW4uY29tL3Jzcw==',
+        spotify_url: 'https://open.spotify.com/show/6fZXOzehJ9JtOyFjirF3qt',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'tonyrobbins',
+        facebook_handle: 'TonyRobbins',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image: 'https://production.listennotes.com/podcasts/the-tony-robbins-podcast-tony-robbins-eh9wNFJcP1W.1400x1400.jpg',
+      title: 'The Tony Robbins Podcast',
+      country: 'United States',
+      website: 'http://tonyrobbins.com/podcast?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [67, 97, 111, 144, 181, 93, 78, 157, 90],
+      itunes_id: 1098413063,
+      publisher: 'Tony Robbins',
+      thumbnail: 'https://production.listennotes.com/podcasts/the-tony-robbins-podcast-tony-robbins-eh9wNFJcP1W.300x300.jpg',
+      is_claimed: false,
+      description:
+        '“Why live an ordinary life, when you can live an extraordinary one?” Tony Robbins, the #1 Life and Business Strategist, has helped over 50 million people from 100 countries create real and lasting change in their lives. In this podcast, he shares proven strategies and tactics so you, too, can achieve massive results in your business, relationships, health and finances. In addition to excerpts from his signature events and other exclusive, never-before-released audio content, Tony and his team also conduct deeply insightful interviews with the most prominent masterminds and experts on the global stage.',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 72,
+      total_episodes: 163,
+      listennotes_url: 'https://www.listennotes.com/c/fc6d33e22b7f4db38df3bb64a9a8c227/',
+      audio_length_sec: 2849,
+      explicit_content: false,
+      latest_episode_id: 'e6472b62839a49589f2bc7c94f4e09da',
+      latest_pub_date_ms: 1694764560000,
+      earliest_pub_date_ms: 1459373820099,
+      update_frequency_hours: 856,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: 'ed79b615ed074204bc4702b56a264a78',
+      rss: 'https://thelifecoachschool.libsyn.com/rss',
+      type: 'episodic',
+      email: 'brooke@thelifecoachschool.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly90aGVsaWZlY29hY2hzY2hvb2wubGlic3luLmNvbS9yc3M=',
+        spotify_url: '',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'brookecastillo',
+        facebook_handle: '',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/the-life-coach-school-podcast-brooke-castillo-xURrlLHEVv--V5of7JlG_RD.1400x1400.jpg',
+      title: 'The Life Coach School Podcast',
+      country: 'United States',
+      website: 'http://www.thelifecoachschool.com?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [67, 88, 90, 93, 94, 111, 115],
+      itunes_id: 870239631,
+      publisher: 'Brooke Castillo',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/the-life-coach-school-podcast-brooke-castillo-vTH8ygOuwXd-V5of7JlG_RD.300x300.jpg',
+      is_claimed: false,
+      description:
+        "The Life Coach School Podcast is your go-to resource for learning, growing, and becoming certified as a Life Coach & Weight Loss Coach. Through this podcast, you'll hear directly from the Master Coach Brooke Castillo to help you better understand life coaching, the required skills and mindsets, and how we focus on serving the client to get them the results they seek.  At The Life Coach School, we offer a thorough and intense certification course that produces some of the most successful coaches coaching today. Learn more at TheLifeCoachSchool.com.",
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 72,
+      total_episodes: 527,
+      listennotes_url: 'https://www.listennotes.com/c/ed79b615ed074204bc4702b56a264a78/',
+      audio_length_sec: 1902,
+      explicit_content: false,
+      latest_episode_id: '651748dadd004534be27e782047599c8',
+      latest_pub_date_ms: 1694682000000,
+      earliest_pub_date_ms: 1398606925484,
+      update_frequency_hours: 168,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: 'bee3a6eeb43d45d482cff16d7e6eec6d',
+      rss: 'https://mcsorleys.barstoolsports.com/feed/bffs',
+      type: 'episodic',
+      email: 'podcasting@barstoolsports.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9tY3NvcmxleXMuYmFyc3Rvb2xzcG9ydHMuY29tL2ZlZWQvYmZmcw==',
+        spotify_url: '',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: '',
+        facebook_handle: '',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/bffs-with-dave-portnoy-josh-richards-and-c-YmGvaLCjt-UwCtO6PxIiq.1400x1400.jpg',
+      title: 'BFFs with Dave Portnoy, Josh Richards, and Brianna Chickenfry',
+      country: 'United States',
+      website:
+        'https://www.barstoolsports.com/shows/134/Bffs?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [93, 98, 67],
+      itunes_id: 1534324153,
+      publisher: 'Barstool Sports',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/bffs-with-dave-portnoy-josh-richards-and-0vXYw5UR7ut-UwCtO6PxIiq.300x300.jpg',
+      is_claimed: false,
+      description:
+        '<p>The unlikely trio of Josh Richards, Dave Portnoy & Brianna Chickenfry team up to talk all things pop culture, celebrities, influencers & TikTok. You never know what to expect from this trio from breaking entertainment news to generational differences they’re sure to make you laugh while keeping you up to date.</p><br /><p>You can find every episode of this show on Apple Podcasts, Spotify or YouTube. Prime Members can listen ad-free on Amazon Music. For more, visit <a href="https://barstool.link/bffspod">barstool.link/bffspod</a></p>',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 71,
+      total_episodes: 144,
+      listennotes_url: 'https://www.listennotes.com/c/bee3a6eeb43d45d482cff16d7e6eec6d/',
+      audio_length_sec: 3716,
+      explicit_content: true,
+      latest_episode_id: 'e97e8c39662c4129b15e09605d7d1984',
+      latest_pub_date_ms: 1694649600000,
+      earliest_pub_date_ms: 1601668154000,
+      update_frequency_hours: 184,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: '2184bada602d431689dbb4c6c1bc5839',
+      rss: 'https://feeds.simplecast.com/atgtihd0',
+      type: 'episodic',
+      email: 'interactive@life.church',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9mZWVkcy5zaW1wbGVjYXN0LmNvbS9hdGd0aWhkMA==',
+        spotify_url: 'https://open.spotify.com/show/7tznexFwtbxfPOYF5mxkxI',
+        youtube_url: 'https://www.youtube.com/@craiggroeschel',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'craiggroeschel',
+        facebook_handle: 'life.church',
+        amazon_music_url: '',
+        instagram_handle: 'craiggroeschel',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/craig-groeschel-leadership-podcast-lifechurch--_K8zgsM0x1-dy-uJsHC_9T.1400x1400.jpg',
+      title: 'Craig Groeschel Leadership Podcast',
+      country: 'United States',
+      website:
+        'https://www.life.church/leadershippodcast?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [93, 97],
+      itunes_id: 1070649025,
+      publisher: 'Life.Church',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/craig-groeschel-leadership-podcast-lifechurch-OU5cY0mgjsb-dy-uJsHC_9T.300x300.jpg',
+      is_claimed: false,
+      description:
+        'The Craig Groeschel Leadership Podcast offers personal, practical coaching lessons that take the mystery out of leadership. In each episode of the Craig Groeschel Leadership Podcast, Craig brings you empowering insights and easy-to-understand takeaways you can use to lead yourself and lead your team. You’ll learn effective ways to grow as a leader, optimize your time, develop your team, and structure your organization.',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 71,
+      total_episodes: 137,
+      listennotes_url: 'https://www.listennotes.com/c/2184bada602d431689dbb4c6c1bc5839/',
+      audio_length_sec: 1770,
+      explicit_content: false,
+      latest_episode_id: '7183c1c09dcb4654aa7d2d02f916482d',
+      latest_pub_date_ms: 1694080800000,
+      earliest_pub_date_ms: 1452675180134,
+      update_frequency_hours: 391,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: 'c5ce6c02cbf1486496206829f7d42e8e',
+      rss: 'https://feeds.megaphone.fm/marketsnacks-daily',
+      type: 'episodic',
+      email: 'iwonder@wondery.com',
+      extra: {
+        url1: 'http://www.marketsnacks.com/',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9mZWVkcy5tZWdhcGhvbmUuZm0vbWFya2V0c25hY2tzLWRhaWx5',
+        spotify_url: 'https://open.spotify.com/show/5RllMBgvDnTau8nnsCUdse',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'marketsnacks',
+        facebook_handle: 'MarketSnacks',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/the-best-one-yet-nick-jack-studios-U_rQkCOW5Zp-kmx0XIZTAys.1400x1400.jpg',
+      title: 'The Best One Yet',
+      country: 'United States',
+      website:
+        'https://wondery.com/shows/the-best-one-yet/?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [98, 67, 93, 95, 99],
+      itunes_id: 1386234384,
+      publisher: 'Nick & Jack Studios',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/the-best-one-yet-nick-jack-studios-Q8pCEFw0caJ-kmx0XIZTAys.300x300.jpg',
+      is_claimed: false,
+      description:
+        '\n      Feel brighter every day with our 20-minute pop-biz podcast. The 3 business news stories you need, with fresh takes you can pretend you came up with — Pairs perfectly with your morning oatmeal ritual. Hosted by Jack Crivici-Kramer & Nick Martell.\n    ',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 71,
+      total_episodes: 1041,
+      listennotes_url: 'https://www.listennotes.com/c/c5ce6c02cbf1486496206829f7d42e8e/',
+      audio_length_sec: 1106,
+      explicit_content: false,
+      latest_episode_id: 'bd76d810f963410daac683fea9dea3f1',
+      latest_pub_date_ms: 1694768400000,
+      earliest_pub_date_ms: 1553519101038,
+      update_frequency_hours: 50,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: 'd863da7f921e435fb35f512b54e774d6',
+      rss: 'https://rss.art19.com/masters-of-scale',
+      type: 'episodic',
+      email: 'hello@mastersofscale.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://play.google.com/music/m/Ify6fvmlw7taa25qkxdricygohe?t=Masters_of_Scale',
+        spotify_url: 'https://open.spotify.com/show/1bJRgaFZHuzifad4IAApFR',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'mastersofscale',
+        facebook_handle: 'mastersofscale',
+        amazon_music_url: '',
+        instagram_handle: 'mastersofscale',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/masters-of-scale-with-reid-hoffman-waitwhat-PJGeHLMmxa6-mYoV0CUyxTD.1400x1400.jpg',
+      title: 'Masters of Scale',
+      country: 'United States',
+      website: 'http://www.mastersofscale.com?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [122, 97, 93, 149, 173, 127, 157, 67, 171],
+      itunes_id: 1227971746,
+      publisher: 'WaitWhat ',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/masters-of-scale-with-reid-hoffman-waitwhat-XJs3WwmUrx7-mYoV0CUyxTD.300x300.jpg',
+      is_claimed: false,
+      description:
+        '\n      <p>Award-winning business advice from Silicon Valley and beyond. Iconic CEOs, from Nike to Netflix, Starbucks to Slack, share the strategies that helped them grow from startups into global brands — and to weather crisis when it strikes. </p><p>Our two formats help tell the complete story of how a business grows, survives and thrives, and the mindsets of growth that keep leaders in the game.</p><p>On each episode of our classic format, host Reid Hoffman — LinkedIn cofounder, Greylock partner and legendary Silicon Valley investor — proves an unconventional theory about how businesses scale, asking his guests to share their stories of entrepreneurship, leadership, strategy, management, fundraising. You’ll hear the human journey too — failures, setbacks, learnings. </p><p>From our Rapid Response format, you can expect real-time wisdom from business leaders in fast-changing situations. Hosted by Bob Safian, past editor in chief of Fast Company, these episodes tackle crisis response, rebuilding, diversity & inclusion, leadership change and much more. </p>\n    ',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 71,
+      total_episodes: 410,
+      listennotes_url: 'https://www.listennotes.com/c/d863da7f921e435fb35f512b54e774d6/',
+      audio_length_sec: 2030,
+      explicit_content: false,
+      latest_episode_id: '0f791bbc7c404dbcb64479e6a94bda25',
+      latest_pub_date_ms: 1694509200000,
+      earliest_pub_date_ms: 1492543297406,
+      update_frequency_hours: 155,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: '5590cb1318054bceb942564a4f067eb6',
+      rss: 'https://www.marketplace.org/feed/podcast/marketplace',
+      type: 'episodic',
+      email: 'marketplacecomments@marketplace.org',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url:
+          'https://podcasts.google.com/feed/aHR0cHM6Ly93d3cubWFya2V0cGxhY2Uub3JnL2ZlZWQvcG9kY2FzdC9tYXJrZXRwbGFjZQ==',
+        spotify_url: 'https://open.spotify.com/show/6zYlX5UGEPmNCWacYUJQGD',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'Marketplace',
+        facebook_handle: 'marketplaceapm',
+        amazon_music_url: '',
+        instagram_handle: 'marketplaceapm',
+      },
+      image: 'https://production.listennotes.com/podcasts/marketplace-marketplace-WHc17NQy23S-Jing2WtK5UE.1400x1400.jpg',
+      title: 'Marketplace',
+      country: 'United States',
+      website:
+        'https://www.marketplace.org/shows/marketplace/?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [95, 99, 173, 93, 67],
+      itunes_id: 201853034,
+      publisher: 'Marketplace',
+      thumbnail: 'https://production.listennotes.com/podcasts/marketplace-marketplace-_YVywHeR0-S-Jing2WtK5UE.300x300.jpg',
+      is_claimed: false,
+      description:
+        '<p>Every weekday, host Kai Ryssdal helps you make sense of the day’s business and economic news — no econ degree or finance background required. “Marketplace” takes you beyond the numbers, bringing you context. Our team of reporters all over the world speak with CEOs, policymakers and regular people just trying to get by.</p>\n',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 71,
+      total_episodes: 453,
+      listennotes_url: 'https://www.listennotes.com/c/5590cb1318054bceb942564a4f067eb6/',
+      audio_length_sec: 1655,
+      explicit_content: false,
+      latest_episode_id: 'ee671f7a3c584b86a97cc57d9d7a6554',
+      latest_pub_date_ms: 1694736272000,
+      earliest_pub_date_ms: 1640304971049,
+      update_frequency_hours: 29,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: '13a7957aeac34e1c9e004f2a2ced5fb0',
+      rss: 'https://feeds.megaphone.fm/pivot',
+      type: 'episodic',
+      email: 'pivot@voxmedia.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9mZWVkcy5tZWdhcGhvbmUuZm0vcGl2b3Q=',
+        spotify_url: 'https://open.spotify.com/show/4MU3RFGELZxPT9XHVwTNPR',
+        youtube_url: '',
+        linkedin_url: 'https://www.linkedin.com/company/re-code-news/',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'recode',
+        facebook_handle: 'RecodeDotNet',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image: 'https://production.listennotes.com/podcasts/pivot-new-york-magazine-b0XsOSV7WDZ-YyJxq4Dbk67.1400x1400.jpg',
+      title: 'Pivot',
+      country: 'United States',
+      website: null,
+      language: 'English',
+      genre_ids: [99, 131, 97, 157, 171, 94, 127, 93, 67],
+      itunes_id: 1073226719,
+      publisher: 'New York Magazine',
+      thumbnail: 'https://production.listennotes.com/podcasts/pivot-new-york-magazine-L2FeZNEqKst-YyJxq4Dbk67.300x300.jpg',
+      is_claimed: false,
+      description:
+        'Every Tuesday and Friday, tech journalist Kara Swisher and NYU Professor Scott Galloway offer sharp, unfiltered insights into the biggest stories in tech, business, and politics. They make bold predictions, pick winners and losers, and bicker and banter like no one else. After all, with great power comes great scrutiny. From New York Magazine and the Vox Media Podcast Network.',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 70,
+      total_episodes: 478,
+      listennotes_url: 'https://www.listennotes.com/c/13a7957aeac34e1c9e004f2a2ced5fb0/',
+      audio_length_sec: 3380,
+      explicit_content: false,
+      latest_episode_id: '14c1faf2af644431a3bb478a290f60e3',
+      latest_pub_date_ms: 1694772000000,
+      earliest_pub_date_ms: 1537812480472,
+      update_frequency_hours: 83,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: '23bd4f3432c2452d93f525e2446a5878',
+      rss: 'https://feeds.simplecast.com/4YELvXgu',
+      type: 'episodic',
+      email: 'rss@earwolf.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9mZWVkcy5zaW1wbGVjYXN0LmNvbS80WUVMdlhndQ==',
+        spotify_url: 'https://open.spotify.com/show/0Vdp4gyQoY0kkcvaLnIgvx',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: '',
+        facebook_handle: '',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/scam-goddess-earwolf-laci-mosley-pMX-87Jicaq-PstEMgqXCUd.1400x1400.jpg',
+      title: 'Scam Goddess',
+      country: 'United States',
+      website: 'https://teamcoco.com/?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [93, 133, 135, 67],
+      itunes_id: 1479455008,
+      publisher: 'Earwolf & Laci Mosley',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/scam-goddess-earwolf-laci-mosley-fy8Bs4RlT06-PstEMgqXCUd.300x300.jpg',
+      is_claimed: false,
+      description:
+        "“Scam Goddess is a podcast dedicated to fraud and all those who practice it! Each week host Laci Mosley (aka Scam Goddess) keeps listeners up to date on current rackets, digs deep into the latest scams, and breaks down historic hoodwinks alongside some of your favorite comedians! It's like true crime only without all the death! True fun crime!”",
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 70,
+      total_episodes: 145,
+      listennotes_url: 'https://www.listennotes.com/c/23bd4f3432c2452d93f525e2446a5878/',
+      audio_length_sec: 4107,
+      explicit_content: true,
+      latest_episode_id: '417a3a3229d4416c8346d5f71afde552',
+      latest_pub_date_ms: 1694491500000,
+      earliest_pub_date_ms: 1420099200036,
+      update_frequency_hours: 168,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: 'd1336f3217eb4510bdbcdf021a38de38',
+      rss: 'https://feeds.megaphone.fm/ADL1740142880',
+      type: 'episodic',
+      email: 'hello@cathyheller.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9mZWVkcy5tZWdhcGhvbmUuZm0vQURMMTc0MDE0Mjg4MA==',
+        spotify_url: 'https://open.spotify.com/show/7DFQyVP0hc3d1xnht1U3UI',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: '',
+        facebook_handle: '',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/the-cathy-heller-podcast-a-podcast-for-sADtGBIXKEc-W-XR2oujCdK.1400x1400.jpg',
+      title: 'The Cathy Heller Podcast: A Podcast for Soulful Entrepreneurs',
+      country: 'United States',
+      website: 'https://www.cathyheller.com/?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [93, 100, 88, 90, 94, 132, 69, 244, 122, 67, 70],
+      itunes_id: 1191831035,
+      publisher: 'Cathy Heller',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/the-cathy-heller-podcast-a-podcast-for-x9FctwM1g6q-W-XR2oujCdK.300x300.jpg',
+      is_claimed: false,
+      description:
+        "Life is too short to be building someone else's dream. On the Cathy Heller Podcast, top ranking podcast host, business and spiritual coach, and bestselling author Cathy Heller gives you the tools to change your frequency, allow in new possibilities, and become the director of your life. You'll learn how to get unstuck from your fears and self doubt, find your confidence, change your subconscious program, shift your mindset, and build an abundant dream life where you can make an impact and get paid to be authentically you. \n\nThis show features coaching calls and heart-to-heart conversations with successful authors, celebrities, entrepreneurs, spiritual leaders, experts, and everyday heroes about what led them to the biggest moments in their lives, and how you too can create a new reality that you can't wait to wake up to every single day. \n\nYou've got so much to contribute and Cathy is here to help you do it. \n\nFollow Cathy @cathy.heller on Instagram for daily sparks of inspiration. \nSubscribe to the Cathy Heller Podcast and share the show with someone who needs to hear it.",
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 70,
+      total_episodes: 791,
+      listennotes_url: 'https://www.listennotes.com/c/d1336f3217eb4510bdbcdf021a38de38/',
+      audio_length_sec: 3065,
+      explicit_content: false,
+      latest_episode_id: '66388e1f91944b9c88c73aa938950317',
+      latest_pub_date_ms: 1694674800000,
+      earliest_pub_date_ms: 1483455806748,
+      update_frequency_hours: 45,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: '499661f3589f42aaa1532673e0e0aedf',
+      rss: 'https://rss.art19.com/smart-passive-income-podcast',
+      type: 'episodic',
+      email: 'podcasts@teamspi.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9yc3MuYXJ0MTkuY29tL3NtYXJ0LXBhc3NpdmUtaW5jb21lLXBvZGNhc3Q=',
+        spotify_url: 'https://open.spotify.com/show/7wjv5MRCXWXImqTFhcufLy',
+        youtube_url: 'https://www.youtube.com/user/smartpassiveincome',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'patflynn',
+        facebook_handle: 'smartpassiveincome',
+        amazon_music_url:
+          'https://music.amazon.com/podcasts/341282a1-cba9-4188-aa58-e09a51ccaa87/the-smart-passive-income-online-business-and-blogging-podcast',
+        instagram_handle: 'patflynn',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/the-smart-passive-income-online-business-jN-aR6qdYuo-NDa6-ySp9kw.1400x1400.jpg',
+      title: 'The Smart Passive Income Online Business and Blogging Podcast',
+      country: 'United States',
+      website:
+        'https://art19.com/shows/smart-passive-income-podcast?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [97, 67, 94, 111, 115, 127, 98, 157, 171, 144, 93, 173],
+      itunes_id: 383084001,
+      publisher: 'Pat Flynn',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/the-smart-passive-income-online-business-sF24owQHYWy-NDa6-ySp9kw.300x300.jpg',
+      is_claimed: false,
+      description:
+        "\n      <p>Pat Flynn from The Smart Passive Income Blog reveals all of his online business and blogging strategies, income sources and killer marketing tips and tricks so you can be ahead of the curve with your online business or blog. Discover how you can create multiple passive income streams that work for you so that you can have the time and freedom to do what you love, whether it's traveling the world, or just living comfortably at home. Since 2008, he's been supporting his family with his many online businesses, and he's been openly sharing his wins, his losses, and all the lessons in between with the community of energetic but humble entrepreneurs who follow him. Self-proclaimed \"crash test dummy of online business\", you'll learn about building authority online, email marketing, building a team and outsourcing, content marketing, podcasting, search engine optimization, niche sites, social media strategies, how to get more traffic, creating online courses, affiliate marketing, and productivity tips so that you create something amazing without burning yourself out. It's a mix of interviews, special co-hosts and solo shows from Pat you're not going to want to miss. Hit subscribe, and get ready to change your life.</p><p><br></p>\n    ",
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 70,
+      total_episodes: 729,
+      listennotes_url: 'https://www.listennotes.com/c/499661f3589f42aaa1532673e0e0aedf/',
+      audio_length_sec: 2471,
+      explicit_content: false,
+      latest_episode_id: '5a7eca894d3d4a33b91a119506b8da9e',
+      latest_pub_date_ms: 1694761200000,
+      earliest_pub_date_ms: 1279551600722,
+      update_frequency_hours: 83,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: '0da3baee05cc47e4b3222da775573efe',
+      rss: 'https://rss.art19.com/the-investors-podcast',
+      type: 'episodic',
+      email: 'stig@theinvestorspodcast.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9yc3MuYXJ0MTkuY29tL3RoZS1pbnZlc3RvcnMtcG9kY2FzdA==',
+        spotify_url: 'https://open.spotify.com/show/28RHOkXkuHuotUrkCdvlOP',
+        youtube_url: 'https://www.youtube.com/user/BuffettsBooks',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'PrestonPysh',
+        facebook_handle: 'WeStudyBillionaires',
+        amazon_music_url:
+          'https://music.amazon.com/podcasts/b42b0171-d75e-4f9a-999f-da3a4db30f45/we-study-billionaires---the-investor%E2%80%99s-podcast-network',
+        instagram_handle: '',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/we-study-billionaires-the-investors-podcast-dufIKDVLEoJ-ZuO23m60ePb.1400x1400.jpg',
+      title: 'We Study Billionaires - The Investor’s Podcast Network',
+      country: 'United States',
+      website: 'https://www.theinvestorspodcast.com?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [97, 144, 93, 111, 67, 98],
+      itunes_id: 928933489,
+      publisher: "The Investor's Podcast Network",
+      thumbnail:
+        'https://production.listennotes.com/podcasts/we-study-billionaires-the-investors-podcast--EpuIXZbUq1-ZuO23m60ePb.300x300.jpg',
+      is_claimed: false,
+      description:
+        '\n      <p>We interview and study famous financial billionaires, including Warren Buffett, Ray Dalio, and Howard Marks, and teach you what we learn and how you can apply their investment strategies in the stock market.</p><p>We Study Billionaires is the largest stock investing podcast show in the world with 100,000,000+ downloads and is hosted by Stig Brodersen and Clay Finck.</p><p>This podcast also includes the Richer Wiser Happier series hosted by best-selling author William Green. William regularly interviews legendary investors such as Mohnish Pabrai and Guy Spier, exploring what they can teach us about how to succeed in markets and life.</p><p>And finally, our Bitcoin Fundamentals series is hosted by Preston Pysh, where he interviews prominent figures in the Bitcoin and macroeconomic space. To learn more about TIP, you can visit <a href="http://theinvestorspodcast.com/" rel="noopener noreferrer" target="_blank">theinvestorspodcast.com</a> or subscribe to our free daily newsletter <a href="https://www.theinvestorspodcast.com/westudymarkets" rel="noopener noreferrer" target="_blank">here</a>.</p>\n    ',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 70,
+      total_episodes: 779,
+      listennotes_url: 'https://www.listennotes.com/c/0da3baee05cc47e4b3222da775573efe/',
+      audio_length_sec: 3439,
+      explicit_content: false,
+      latest_episode_id: '3ed912d88ff440799e03b2e847cbe32c',
+      latest_pub_date_ms: 1694736000000,
+      earliest_pub_date_ms: 1411981990769,
+      update_frequency_hours: 57,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: 'f4a7bdbef7a84fd0b4a712b70a3c5ec5',
+      rss: 'https://choosefi.libsyn.com/rss',
+      type: 'episodic',
+      email: 'feedback@choosefi.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9jaG9vc2VmaS5saWJzeW4uY29tL3Jzcw==',
+        spotify_url: '',
+        youtube_url: 'https://www.youtube.com/choosefi',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'choosefi',
+        facebook_handle: 'ChooseFi',
+        amazon_music_url: '',
+        instagram_handle: 'choosefiradio',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/choosefi-jonathan-mendonsa-brad-barrett-Xb8c9pQA_y--e8_g1GAYHIj.1400x1400.jpg',
+      title: 'ChooseFI',
+      country: 'United States',
+      website: 'https://www.choosefi.com?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [67, 93, 97, 122, 123, 127, 129, 144, 98],
+      itunes_id: 1187770032,
+      publisher: 'ChooseFI',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/choosefi-jonathan-mendonsa-brad-barrett-UCfs8WRklv0-e8_g1GAYHIj.300x300.jpg',
+      is_claimed: false,
+      description:
+        'How would your life change if you reached Financial Independence and got to the point where working is optional? What actions can you take today to make that not just possible but probable. Jonathan & Brad explore the tactics that the FI community uses to reclaim decades of their lives. They discuss reducing expenses, crushing debt, tax optimization, building passive income streams through online businesses and real estate and how to travel the world for free. Every episode is packed with actionable tips and no topic is too big or small as long as it speeds up the process of reaching financial independence.',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 69,
+      total_episodes: 615,
+      listennotes_url: 'https://www.listennotes.com/c/f4a7bdbef7a84fd0b4a712b70a3c5ec5/',
+      audio_length_sec: 3359,
+      explicit_content: false,
+      latest_episode_id: '5aa1535467ee46aa9f183f5a92069135',
+      latest_pub_date_ms: 1694408400000,
+      earliest_pub_date_ms: 1482062785609,
+      update_frequency_hours: 167,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: 'c73271d55ffa4e2d9b529220072fbd79',
+      rss: 'https://www.omnycontent.com/d/playlist/e73c998e-6e60-432f-8610-ae210140c5b1/3e7f11c6-1170-40a0-be58-ae330037e2f5/dea7cb4c-cbee-49ce-83f0-ae330037e303/podcast.rss',
+      type: 'episodic',
+      email: 'michael@earnyourleisure.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url:
+          'https://podcasts.google.com/feed/aHR0cHM6Ly93d3cub21ueWNvbnRlbnQuY29tL2QvcGxheWxpc3QvZTczYzk5OGUtNmU2MC00MzJmLTg2MTAtYWUyMTAxNDBjNWIxLzNlN2YxMWM2LTExNzAtNDBhMC1iZTU4LWFlMzMwMDM3ZTJmNS9kZWE3Y2I0Yy1jYmVlLTQ5Y2UtODNmMC1hZTMzMDAzN2UzMDMvcG9kY2FzdC5yc3M=',
+        spotify_url: 'https://open.spotify.com/show/2S4tSSlT71Z5i8Dt1vlDJc',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: '',
+        facebook_handle: '',
+        amazon_music_url: '',
+        instagram_handle: '',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/earn-your-leisure-eyl-network--AFWvtrYomD-CSRy4Lz625Y.1400x1400.jpg',
+      title: 'Earn Your Leisure',
+      country: 'United States',
+      website:
+        'https://redcircle.com/shows/earn-your-leisure7962?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [67, 93],
+      itunes_id: 1450211392,
+      publisher: 'EYL Network',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/earn-your-leisure-eyl-network-Ni5dRWAQFJc-CSRy4Lz625Y.300x300.jpg',
+      is_claimed: false,
+      description:
+        '<p>Welcome to The Earn Your Leisure Podcast. Rashad Bilal and Troy Millings will be your host. Earn Your Leisure will be giving you behind the scenes financial views into the entertainment and sports industries as well as highlighting back stories of entrepreneurs. We will also be breaking down business models and examining the latest trends in finance. Earn Your Leisure is a college business class mixed with pop culture. We blend the two together for a unique and exciting look into the world of business. Let’s go!! #earnyourleisurepodcast</p>',
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 69,
+      total_episodes: 743,
+      listennotes_url: 'https://www.listennotes.com/c/c73271d55ffa4e2d9b529220072fbd79/',
+      audio_length_sec: 2289,
+      explicit_content: false,
+      latest_episode_id: '190fb85692fd4afbb1b690b9cbead741',
+      latest_pub_date_ms: 1694782800000,
+      earliest_pub_date_ms: 1548186120730,
+      update_frequency_hours: 23,
+      listen_score_global_rank: '0.05%',
+    },
+    {
+      id: 'fbecfdd4116e4e7a954bd6bc4cb0b406',
+      rss: 'https://amyporterfield.libsyn.com/rss',
+      type: 'episodic',
+      email: 'podcast@amyporterfield.com',
+      extra: {
+        url1: '',
+        url2: '',
+        url3: '',
+        google_url: 'https://podcasts.google.com/feed/aHR0cHM6Ly9hbXlwb3J0ZXJmaWVsZC5saWJzeW4uY29tL3Jzcw==',
+        spotify_url: 'https://open.spotify.com/show/5z7TqC6tll8egI9prMqXhd',
+        youtube_url: '',
+        linkedin_url: '',
+        wechat_handle: '',
+        patreon_handle: '',
+        twitter_handle: 'AmyPorterfield',
+        facebook_handle: 'AmyPorterfield',
+        amazon_music_url: '',
+        instagram_handle: 'amyporterfield',
+      },
+      image:
+        'https://production.listennotes.com/podcasts/online-marketing-made-easy-with-amy-HmKNg8E7VI0-jXUyf4vBV20.1400x1400.jpg',
+      title: 'Online Marketing Made Easy with Amy Porterfield',
+      country: 'United States',
+      website:
+        'https://amyporterfield.com/category/podcast/?utm_source=listennotes.com&utm_campaign=Listen+Notes&utm_medium=website',
+      language: 'English',
+      genre_ids: [67, 97, 94, 157, 93, 173],
+      itunes_id: 594703545,
+      publisher: 'Amy Porterfield',
+      thumbnail:
+        'https://production.listennotes.com/podcasts/online-marketing-made-easy-with-amy-yMjTufCWtoa-jXUyf4vBV20.300x300.jpg',
+      is_claimed: false,
+      description:
+        "Ever wish you had a business mentor with over a decade of experience whispering success secrets in your ear? That’s exactly what you’ll get when you tune into the top-ranked Online Marketing Made Easy Podcast with your host, 9 to 5er turned CEO of a multi-million dollar business & NY Times Best Selling Author, Amy Porterfield. Her specialty? Breaking down big ideas and strategies into actionable step-by-step processes designed to get you results with a whole lot less stress. Tune in, get inspired, and get ready to discover why hundreds of thousands of online business owners turn to Amy for guidance when it comes to all things online business including digital courses, list building, social media, content, webinars, and so much more. Whether you're a budding entrepreneur, have a comfy side-hustle, or are looking to take your business to the next level, each episode is designed to help you take immediate action on the most important strategies for starting and growing your online business today.",
+      looking_for: {
+        guests: false,
+        cohosts: false,
+        sponsors: false,
+        cross_promotion: false,
+      },
+      listen_score: 69,
+      total_episodes: 652,
+      listennotes_url: 'https://www.listennotes.com/c/fbecfdd4116e4e7a954bd6bc4cb0b406/',
+      audio_length_sec: 2257,
+      explicit_content: false,
+      latest_episode_id: '7d5d4a8d3fbe4518acaf27937749ac22',
+      latest_pub_date_ms: 1694674860000,
+      earliest_pub_date_ms: 1358200867624,
+      update_frequency_hours: 73,
+      listen_score_global_rank: '0.05%',
+    },
+  ],
+  parent_id: 67,
+  page_number: 2,
+  has_previous: true,
+  listennotes_url: 'https://www.listennotes.com/best-business-podcasts-93/',
+  next_page_number: 3,
+  previous_page_number: 1,
+};
